+++ vuepress/shared/changelog.md
@@ -4,6 +4,9 @@
 
 ### v3.0.1
   - bugfix: header title and aria label incorrectly informs user that a column can be sorted when sort is turned on for the grid but not that specific column
+  - accessibility: scrollable region must have keyboard access for dgv-data-grid-body
+  - accessibility: aria-label attribute cannot be used on a span with no valid role attribute dgv-filter-operator-role
+
 
 ### v3.0.0
   ::: danger BREAKING